+++ src/components/Dashboard/Sidebar.jsx
@@ -2,13 +2,10 @@ import {
     Logo,
     Menu,
     reactLogo,
-    useState,
     ADMIN_MODE,
 } from '../barrel_module/Barrel.jsx';
 
-function Sidebar() {
-    const [activeMenu, setActiveMenu] = useState("Dashboard");
-
+function Sidebar({ activeMenu, setActiveMenu }) {
     const menuNames = [];
     
     if(ADMIN_MODE) {
@@ -26,14 +23,14 @@ function Sidebar() {
 
     return (
         <div className="hidden lg:flex lg:flex-col lg:border-e-1 lg:border-gray-300 lg:min-w-[22rem] lg:visible">
-            <Logo textsize={"text-4xl"} margin={"mb-5"}/>
+            <Logo textsize={"text-4xl"} margin={"mb-8"}/>
             {menuNames.map(menu => (
                 <Menu 
                     key={menu.key}
                     menuName={menu.name} 
                     iconUrl={menu.iconUrl} 
-                    isActive={activeMenu === menu.name} 
-                    setActive={() => setActiveMenu(menu.name)}/>
+                    isActive={activeMenu === menu.key} 
+                    setActive={() => setActiveMenu(menu.key)}/>
             ))}
         </div>
     )